requirements.txt
fastapi[all]
sqlalchemy
sqlmodel
alembic
psycopg2-binary
python-dotenv
uvicorn
pika==1.3.2
pydantic
python-telegram-bot
prometheus-client
python-telegram-bot>=20.0

pytest
pytest-cov
pytest-asyncio
pytest-env
httpx
async-timeout
pytest-httpx

docker>=7.1.0
pytest-docker>=3.2.0

prometheus-client
